+++ src/main/kotlin/cloud/glitchdev/rfu/manager/mob/MobManager.kt
@@ -45,12 +45,13 @@ object MobManager : RegisteredEvent {
     private fun scanEntities(world: ClientLevel) {
         var foundAnyFlare = false
         world.entitiesForRendering().forEach { entity ->
-            if (entity !is ArmorStand) return@forEach
+            if (entity is ArmorStand) {
+                checkSbEntity(entity, world)
+            }
 
             if (DeployableManager.checkEntity(entity)) {
                 foundAnyFlare = true
             }
-            checkSbEntity(entity, world)
         }
 
         if (!foundAnyFlare) {